requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
chardet==7.1.0
charset-normalizer==3.4.5
click==7.1.2
colorama==0.4.6
Flask==2.0.3
idna==3.11
interchange==2021.0.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
monotonic==1.6
nltk==3.9.3
packaging==26.0
pansi==2024.11.0
pillow==12.1.1
py2neo==2021.2.4
Pygments==2.19.2
python-dotenv==1.2.2
pytz==2026.1.post1
regex==2026.2.28
requests==2.32.5
six==1.17.0
slacker==0.14.0
soupsieve==2.8.3
tqdm==4.67.3
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.6.3
Werkzeug==2.0.3
